setup.py: replace composer with mosaicml (#458)
When we build mosaicml with "pip install -e .", we'll see package
mosaicml is installed with link to the code tree of composer.

Without this change, we'll see package composer is again installed
without link to the code tree of composer. mosaicml and composer
are two different packages and llm-foundry will call into package
composer which is not what we expect.

With this change, llm-foundry will call into the code tree of
composer which is our expectation.